Move createNewApiKey to within integrations
